Refactor gradle plugin task
